add the service registration
