--- src/aliases/api/handlers/aliases.go
@@ -1,11 +1,0 @@
-package handlers
-
-//type AliasHandler struct {
-//	alias aliases.Alias
-//}
-//
-//func NewAliasHandler(backend aliases.Alias) *AliasHandler {
-//	return &AliasHandler{
-//		alias: backend,
-//	}
-//}
